Added style guide and casting command.
Ah, Eclipse was being slow, and I have failed in the attempt to speed
it up. I was tempted to bump this release up to V0.7, given all the
effort required for this relatively small change, and consequently
introduce a new test image, but that will have to wait until $:compile.
Anyway, read the freaking style guide, and, Eclipse, give me a freaking
break and let me just program. Please?